+++ device/tt_device/tt_device.cpp
@@ -15,6 +15,7 @@ TTDevice::TTDevice(
     std::unique_ptr<PCIDevice> pci_device, std::unique_ptr<architecture_implementation> architecture_impl) :
     pci_device_(std::move(pci_device)),
     architecture_impl_(std::move(architecture_impl)),
+    tlb_manager_(std::make_unique<TLBManager>(this)),
     arch(architecture_impl_->get_architecture()) {}
 
 /* static */ std::unique_ptr<TTDevice> TTDevice::create(int pci_device_number) {
@@ -35,6 +36,8 @@ TTDevice::TTDevice(
 architecture_implementation *TTDevice::get_architecture_implementation() { return architecture_impl_.get(); }
 
 PCIDevice *TTDevice::get_pci_device() { return pci_device_.get(); }
+
+TLBManager *TTDevice::get_tlb_manager() { return tlb_manager_.get(); }
 
 bool TTDevice::is_hardware_hung() {
     volatile const void *addr = reinterpret_cast<const char *>(pci_device_->bar0_uc) +
